refactor: remove CardHtml
Replaced with `RenderedCard`

All methods moved into HtmlGenerator
