+++ AnkiDroid/src/main/java/com/ichi2/anki/cardviewer/HtmlGenerator.kt
@@ -17,7 +17,6 @@
 package com.ichi2.anki.cardviewer
 
 import android.content.Context
-import android.content.res.Resources
 import androidx.annotation.CheckResult
 import anki.config.ConfigKey
 import com.ichi2.anki.preferences.sharedPrefs
@@ -26,28 +25,83 @@ import com.ichi2.libanki.Card
 import com.ichi2.libanki.Collection
 import com.ichi2.libanki.Sound
 import com.ichi2.libanki.stripAvRefs
+import com.ichi2.libanki.template.MathJax
+import timber.log.Timber
 
+/**
+ * Holds Android-specific context which affects how a card is rendered to HTML
+ *
+ * @see generateHtml
+ */
 class HtmlGenerator(
     private val typeAnswer: TypeAnswer,
-    val cardAppearance: CardAppearance,
-    val cardTemplate: CardTemplate,
-    private val showAudioPlayButtons: Boolean,
-    val resources: Resources
+    private val cardAppearance: CardAppearance,
+    private val cardTemplate: CardTemplate,
+    private val showAudioPlayButtons: Boolean
 ) {
 
+    /**
+     * Renders Android-specific functionality to produce a [RenderedCard]
+     */
     @CheckResult
-    fun generateHtml(col: Collection, card: Card, side: SingleCardSide): CardHtml {
-        return CardHtml.createInstance(col, card, side, this)
+    fun generateHtml(col: Collection, card: Card, side: SingleCardSide): RenderedCard {
+        // obtain the libAnki-rendered card
+        var content: String = if (side == SingleCardSide.FRONT) card.question(col) else card.answer(col)
+        // IRI-encodes media: `foo bar` -> `foo%20bar`
+        content = col.media.escapeMediaFilenames(content)
+        // produces either an <input> or <span>...</span> to denote typed input
+        content = filterTypeAnswer(content, side)
+        // wraps content in <div id="qa">
+        content = enrichWithQADiv(content)
+        // expands [anki:q:1] to a play button
+        content = expandSounds(content)
+        // fixes an Android bug where font-weight:600 does not display
+        content = CardAppearance.fixBoldStyle(content)
+
+        // based on the content, load appropriate scripts such as MathJax, then render
+        return render(content, card.ord)
     }
 
-    fun filterTypeAnswer(content: String, side: SingleCardSide): String {
+    private fun render(content: String, ord: Int): RenderedCard {
+        val requiresMathjax = MathJax.textContainsMathjax(content)
+
+        val style = cardAppearance.style
+        val script = when (requiresMathjax) {
+            false -> ""
+            true ->
+                """        <script src="file:///android_asset/mathjax/conf.js"></script>
+        <script src="file:///android_asset/mathjax/tex-chtml.js"></script>"""
+        }
+        val cardClass = cardAppearance.getCardClass(ord + 1) + if (requiresMathjax) " mathjax-needs-to-render" else ""
+
+        Timber.v("content card = \n %s", content)
+        Timber.v("::style:: / %s", style)
+
+        return cardTemplate.render(content, style, script, cardClass)
+    }
+
+    /**
+     * Adds a div html tag around the contents to have an indication, where answer/question is displayed
+     *
+     * @param content The content to surround with tags.
+     * @return The enriched content
+     */
+    private fun enrichWithQADiv(content: String?): String {
+        val sb = StringBuilder()
+        sb.append("""<div id="qa">""")
+        sb.append(content)
+        sb.append("</div>")
+        return sb.toString()
+    }
+
+    private fun filterTypeAnswer(content: String, side: SingleCardSide): String {
         return when (side) {
             SingleCardSide.FRONT -> typeAnswer.filterQuestion(content)
             SingleCardSide.BACK -> typeAnswer.filterAnswer(content)
         }
     }
 
-    fun expandSounds(content: String): String {
+    private fun expandSounds(content: String): String {
         return if (showAudioPlayButtons) {
             Sound.expandSounds(content)
         } else {
@@ -69,8 +123,7 @@ class HtmlGenerator(
                 typeAnswer,
                 cardAppearance,
                 cardHtmlTemplate,
-                showAudioPlayButtons,
-                context.resources
+                showAudioPlayButtons
             )
         }
     }
